+++ doc/api/pymongo/collection.rst
@@ -67,6 +67,4 @@
       .. automethod:: options
       .. automethod:: map_reduce
       .. automethod:: inline_map_reduce
-      .. automethod:: initialize_unordered_bulk_op
-      .. automethod:: initialize_ordered_bulk_op
       .. automethod:: count
