+++ pymongo/bulk.py
@@ -521,183 +521,3 @@ class _Bulk(object):
                 self.execute_no_results(sock_info, generator)
         else:
             return self.execute_command(generator, write_concern, session)
-
-
-class BulkUpsertOperation(object):
-    """An interface for adding upsert operations.
-    """
-
-    __slots__ = ('__selector', '__bulk', '__collation')
-
-    def __init__(self, selector, bulk, collation):
-        self.__selector = selector
-        self.__bulk = bulk
-        self.__collation = collation
-
-    def update_one(self, update):
-        """Update one document matching the selector.
-
-        :Parameters:
-          - `update` (dict): the update operations to apply
-        """
-        self.__bulk.add_update(self.__selector,
-                               update, multi=False, upsert=True,
-                               collation=self.__collation)
-
-    def update(self, update):
-        """Update all documents matching the selector.
-
-        :Parameters:
-          - `update` (dict): the update operations to apply
-        """
-        self.__bulk.add_update(self.__selector,
-                               update, multi=True, upsert=True,
-                               collation=self.__collation)
-
-    def replace_one(self, replacement):
-        """Replace one entire document matching the selector criteria.
-
-        :Parameters:
-          - `replacement` (dict): the replacement document
-        """
-        self.__bulk.add_replace(self.__selector, replacement, upsert=True,
-                                collation=self.__collation)
-
-
-class BulkWriteOperation(object):
-    """An interface for adding update or remove operations.
-    """
-
-    __slots__ = ('__selector', '__bulk', '__collation')
-
-    def __init__(self, selector, bulk, collation):
-        self.__selector = selector
-        self.__bulk = bulk
-        self.__collation = collation
-
-    def update_one(self, update):
-        """Update one document matching the selector criteria.
-
-        :Parameters:
-          - `update` (dict): the update operations to apply
-        """
-        self.__bulk.add_update(self.__selector, update, multi=False,
-                               collation=self.__collation)
-
-    def update(self, update):
-        """Update all documents matching the selector criteria.
-
-        :Parameters:
-          - `update` (dict): the update operations to apply
-        """
-        self.__bulk.add_update(self.__selector, update, multi=True,
-                               collation=self.__collation)
-
-    def replace_one(self, replacement):
-        """Replace one entire document matching the selector criteria.
-
-        :Parameters:
-          - `replacement` (dict): the replacement document
-        """
-        self.__bulk.add_replace(self.__selector, replacement,
-                                collation=self.__collation)
-
-    def remove_one(self):
-        """Remove a single document matching the selector criteria.
-        """
-        self.__bulk.add_delete(self.__selector, _DELETE_ONE,
-                               collation=self.__collation)
-
-    def remove(self):
-        """Remove all documents matching the selector criteria.
-        """
-        self.__bulk.add_delete(self.__selector, _DELETE_ALL,
-                               collation=self.__collation)
-
-    def upsert(self):
-        """Specify that all chained update operations should be
-        upserts.
-
-        :Returns:
-          - A :class:`BulkUpsertOperation` instance, used to add
-            update operations to this bulk operation.
-        """
-        return BulkUpsertOperation(self.__selector, self.__bulk,
-                                   self.__collation)
-
-
-class BulkOperationBuilder(object):
-    """**DEPRECATED**: An interface for executing a batch of write operations.
-    """
-
-    __slots__ = '__bulk'
-
-    def __init__(self, collection, ordered=True,
-                 bypass_document_validation=False):
-        """**DEPRECATED**: Initialize a new BulkOperationBuilder instance.
-
-        :Parameters:
-          - `collection`: A :class:`~pymongo.collection.Collection` instance.
-          - `ordered` (optional): If ``True`` all operations will be executed
-            serially, in the order provided, and the entire execution will
-            abort on the first error. If ``False`` operations will be executed
-            in arbitrary order (possibly in parallel on the server), reporting
-            any errors that occurred after attempting all operations. Defaults
-            to ``True``.
-          - `bypass_document_validation`: (optional) If ``True``, allows the
-            write to opt-out of document level validation. Default is
-            ``False``.
-
-        .. note:: `bypass_document_validation` requires server version
-          **>= 3.2**
-
-        .. versionchanged:: 3.5
-           Deprecated. Use :meth:`~pymongo.collection.Collection.bulk_write`
-           instead.
-
-        .. versionchanged:: 3.2
-          Added bypass_document_validation support
-        """
-        self.__bulk = _Bulk(collection, ordered, bypass_document_validation)
-
-    def find(self, selector, collation=None):
-        """Specify selection criteria for bulk operations.
-
-        :Parameters:
-          - `selector` (dict): the selection criteria for update
-            and remove operations.
-          - `collation` (optional): An instance of
-            :class:`~pymongo.collation.Collation`. This option is only
-            supported on MongoDB 3.4 and above.
-
-        :Returns:
-          - A :class:`BulkWriteOperation` instance, used to add
-            update and remove operations to this bulk operation.
-
-        .. versionchanged:: 3.4
-           Added the `collation` option.
-
-        """
-        validate_is_mapping("selector", selector)
-        return BulkWriteOperation(selector, self.__bulk, collation)
-
-    def insert(self, document):
-        """Insert a single document.
-
-        :Parameters:
-          - `document` (dict): the document to insert
-
-        .. seealso:: :ref:`writes-and-ids`
-        """
-        self.__bulk.add_insert(document)
-
-    def execute(self, write_concern=None):
-        """Execute all provided operations.
-
-        :Parameters:
-          - write_concern (optional): the write concern for this bulk
-            execution.
-        """
-        if write_concern is not None:
-            write_concern = WriteConcern(**write_concern)
-        return self.__bulk.execute(write_concern, session=None)
